% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountries.R
\name{getCountries}
\alias{getCountries}
\title{Get country names from YouTheria}
\usage{
getCountries()
}
\value{
A dataframe of country names and IDs. These names can be used in 
\link{getMeasurementData} to restrict the search to a specific country
}
\description{
Retrieves a \code{data.frame} of country names and IDs from YouTheria.
}
\examples{
\dontrun{
# Get a dataframe of all countries
getCountries()
}
}

