\name{rYoutheria}
\alias{rYoutheria}
\title{The rYoutheria package}
\description{
  The rYouTheria package harnesses the YouTheria wedsite's APIs to give easy access
  to a large amount of mammal trait data. The package allows the user to select data 
  by species, location and measurement type. More information can be found out about
  the YouTheria dataset at \url{http://www.utheria.org/}
  \cr
  \cr
  Click on the index link below for more information on each 
  function  
}
