\name{getCountries}
\alias{getCountries}
\title{Get country names from YouTheria}
\usage{
getCountries()
}
\value{
A dataframe of country names and IDs. These names can be
used in \link{getMeasurementData} to restrict the search to
a specific country
}
\description{
Retrieves a \code{data.frame} of country names and IDs from
Youtheria.
}
\examples{
\dontrun{

getCountries()

}
}

